+++ src/components/projects/ProjectsList.js
@@ -4,22 +4,21 @@ import clsx from 'clsx';
 
 import { makeStyles } from '@material-ui/core';
 
-import { ProjectsListItem } from 'custom-components';
+import { Loader, ProjectsListItem } from 'custom-components';
 
 const useStyles = makeStyles((theme) => ({
     root: {}
 }));
 
-const ProjectsList = ({ classes, items, ...props }) => {
+const ProjectsList = ({ classes, items, loading, ...props }) => {
     const internalClasses = useStyles();
 
     // const textProvider = useSelector(({ ui }) => ui.textContent);
 
     return (
         <div className={clsx(internalClasses.root, classes?.root)} {...props}>
-            {items?.map((item, index) => (
-                <ProjectsListItem key={item.id || index} item={item} />
-            ))}
+            <Loader loading={loading} />
+            {!loading && items?.map((item, index) => <ProjectsListItem key={item.id || index} item={item} />)}
         </div>
     );
 };