+++ src/store/projects/currentProjectSlice.js
@@ -1,42 +1,45 @@
-import { createSlice } from '@reduxjs/toolkit';
+import { createAsyncThunk, createEntityAdapter, createSlice } from '@reduxjs/toolkit';
 import firebaseService from 'services/firebaseService';
 
-export const loadCurrentProject = (id) => (dispatch, getState) => {
-    dispatch(projectRequested());
-    return firebaseService
-        .getProject(id)
-        .then((data) => {
-            dispatch(projectReceived(data));
-        })
-        .catch((error) => {
-            dispatch(projectRequestFailed(error));
-        });
-};
+export const loadCurrentProject = createAsyncThunk(
+    'currentProject/loadCurrentProject',
+    async (id, { dispatch, getState }) => {
+        try {
+            const data = await firebaseService.getProject(id);
+            return data;
+        } catch (error) {
+            return `${error?.message}`;
+        }
+    }
+);
+
+const currentProjectAdapter = createEntityAdapter({});
+
+export const { selectAll: selectCurrentProjects } = currentProjectAdapter.getSelectors(
+    ({ entities }) => entities.projects.currentProject
+);
 
 const slice = createSlice({
     name: 'currentProject',
     initialState: {
         data: {},
+        entities: {},
+        ids: [],
         lastFetch: '',
         loading: false
     },
-    reducers: {
-        projectRequested: (state, action) => {
+    reducers: {},
+    extraReducers: {
+        [loadCurrentProject.pending]: (state, action) => {
             state.loading = true;
         },
-
-        projectReceived: (state, action) => {
-            state.data = action.payload;
-            state.loading = false;
-            state.lastFetch = Date.now();
-        },
-
-        projectRequestFailed: (state, action) => {
+        [loadCurrentProject.fulfilled]: (state, action) => {
+            currentProjectAdapter.setAll(state, action.payload);
             state.loading = false;
         }
     }
 });
 
-export const { projectReceived, projectRequestFailed, projectRequested } = slice.actions;
+// export const {} = slice.actions;
 
 export default slice.reducer;