+++ src/store/projects/userProjectsSlice.js
@@ -1,54 +1,54 @@
-import { createSlice } from '@reduxjs/toolkit';
+import { createAsyncThunk, createEntityAdapter, createSlice } from '@reduxjs/toolkit';
 import firebaseService from 'services/firebaseService';
 
-export const loadUserProjects = () => (dispatch, getState) => {
-    dispatch(projectsRequested());
-    return firebaseService
-        .getUserProjects()
-        .then((data) => {
-            dispatch(projectsReceived(data));
-        })
-        .catch((error) => {
-            dispatch(projectsRequestFailed(error));
-        });
-};
+export const loadUserProjects = createAsyncThunk(
+    'userProjects/loadUserProjects',
+    async (params, { dispatch, getState }) => {
+        try {
+            const data = await firebaseService.getUserProjects(params);
+            return data;
+        } catch (error) {
+            return `${error?.message}`;
+        }
+    }
+);
+
+const userProjectsAdapter = createEntityAdapter({});
+
+export const { selectAll: selectUserProjects, selectById: selectUserProjectsById } = userProjectsAdapter.getSelectors(
+    ({ entities }) => entities.projects.userProjects
+);
 
 const slice = createSlice({
-    name: 'userProjects',
+    name: 'projects/userProjects',
     initialState: {
+        entities: {},
+        ids: [],
+        loading: false,
         searchParams: {
             searchText: ''
-        },
-        list: [],
-        loading: false,
-        lastFetch: null
+        }
     },
     reducers: {
-        projectsRequested: (data, action) => {
-            data.loading = true;
+        setSearchText: (state, action) => {
+            state.searchParams.searchText = action.payload;
         },
 
-        projectsReceived: (data, action) => {
-            data.list = action.payload;
-            data.loading = false;
-            data.lastFetch = Date.now();
+        clearSearchParams: (state, action) => {
+            state.searchParams.searchText = '';
+        }
+    },
+    extraReducers: {
+        [loadUserProjects.pending]: (state, action) => {
+            state.loading = true;
         },
-
-        projectsRequestFailed: (data, action) => {
-            data.loading = false;
-        },
-
-        setSearchText: (data, action) => {
-            data.searchParams.searchText = action.payload;
-        },
-
-        clearSearchParams: (data, action) => {
-            data.searchParams.searchText = '';
+        [loadUserProjects.fulfilled]: (state, action) => {
+            userProjectsAdapter.setAll(state, action.payload);
+            state.loading = false;
         }
     }
 });
 
-export const { projectsReceived, projectsRequestFailed, projectsRequested, setSearchText, clearSearchParams } =
-    slice.actions;
+export const { setProjects, setSearchText, clearSearchParams } = slice.actions;
 
 export default slice.reducer;